chore(volo-http): separate client & server, split features
