+++ examples/Cargo.toml
@@ -92,6 +92,8 @@ anyhow.workspace = true
 async-stream.workspace = true
 bytes.workspace = true
 faststr.workspace = true
+http.workspace = true
+http-body.workspace = true
 http-body-util.workspace = true
 lazy_static.workspace = true
 metainfo.workspace = true
