+++ volo-http/src/context/mod.rs
@@ -1,51 +1,3 @@
-macro_rules! impl_deref_and_deref_mut {
-    ($type:ty, $inner:ty, $pos:tt) => {
-        impl std::ops::Deref for $type {
-            type Target = $inner;
-
-            #[inline]
-            fn deref(&self) -> &Self::Target {
-                &self.$pos
-            }
-        }
-
-        impl std::ops::DerefMut for $type {
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut self.$pos
-            }
-        }
-    };
-}
-
-macro_rules! impl_getter {
-    ($name: ident, $type: ty, $($path: tt).+) => {
-        paste! {
-            #[inline]
-            pub fn $name(&self) -> &$type {
-                &self.$($path).+
-            }
-
-            #[inline]
-            pub fn [<$name _mut>](&mut self) -> &mut $type {
-                &mut self.$($path).+
-            }
-        }
-    };
-    ($name: ident, $type: ty) => {
-        paste! {
-            #[inline]
-            pub fn $name(&self) -> &$type {
-                &self.$name
-            }
-
-            #[inline]
-            pub fn [<$name _mut>](&mut self) -> &mut $type {
-                &mut self.$name
-            }
-        }
-    };
-}
-
 macro_rules! stat_impl {
     ($t: ident) => {
         paste! {
@@ -71,15 +23,15 @@ macro_rules! stat_impl {
     };
 }
 
+#[cfg(feature = "client")]
 pub mod client;
+#[cfg(feature = "client")]
+pub use self::client::ClientContext;
+
+#[cfg(feature = "server")]
 pub mod server;
-
-pub use self::{
-    client::ClientContext,
-    server::{ConnectionInfo, ServerContext},
-};
-
-pub type HttpContext = ServerContext;
+#[cfg(feature = "server")]
+pub use self::server::{ConnectionInfo, ServerContext};
 
 /// This is unstable now and may be changed in the future.
 #[derive(Debug, Default, Clone, Copy)]
