+++ volo-http/src/json/server.rs
@@ -5,60 +5,17 @@ use http::{
     StatusCode,
 };
 use hyper::body::Incoming;
-use serde::{de::DeserializeOwned, ser::Serialize};
-#[cfg(all(feature = "serde_json", feature = "sonic_json"))]
-compile_error!("features `serde_json` and `sonic_json` cannot be enabled at the same time.");
-#[cfg(feature = "serde_json")]
-pub use serde_json::Error;
-#[cfg(feature = "sonic_json")]
-pub use sonic_rs::Error;
+use serde::de::DeserializeOwned;
 
+use super::{deserialize, Error, Json};
 use crate::{
-    body::Body,
     context::ServerContext,
-    extract::{FromRequest, RejectionError},
-    response::{IntoResponse, ServerResponse},
+    response::ServerResponse,
+    server::{
+        extract::{FromRequest, RejectionError},
+        IntoResponse,
+    },
 };
-
-pub(crate) fn serialize<T>(data: &T) -> Result<Vec<u8>, Error>
-where
-    T: Serialize,
-{
-    #[cfg(feature = "sonic_json")]
-    let res = sonic_rs::to_vec(data);
-
-    #[cfg(feature = "serde_json")]
-    let res = serde_json::to_vec(data);
-
-    res
-}
-
-pub(crate) fn deserialize<T>(data: &[u8]) -> Result<T, Error>
-where
-    T: DeserializeOwned,
-{
-    #[cfg(feature = "sonic_json")]
-    let res = sonic_rs::from_slice(data);
-
-    #[cfg(feature = "serde_json")]
-    let res = serde_json::from_slice(data);
-
-    res
-}
-
-#[derive(Debug, Default, Clone, Copy)]
-pub struct Json<T>(pub T);
-
-impl<T> TryFrom<Json<T>> for Body
-where
-    T: Serialize,
-{
-    type Error = Error;
-
-    fn try_from(value: Json<T>) -> Result<Self, Self::Error> {
-        serialize(&value.0).map(Body::from)
-    }
-}
 
 impl IntoResponse for Error {
     fn into_response(self) -> ServerResponse {
