+++ volo-http/src/lib.rs
@@ -1,25 +1,19 @@
 pub mod body;
+#[cfg(feature = "client")]
 pub mod client;
 pub mod context;
 #[cfg(feature = "cookie")]
 pub mod cookie;
 pub mod error;
 pub mod extension;
-pub mod extract;
-pub mod handler;
-#[cfg(any(feature = "serde_json", feature = "sonic_json"))]
+#[cfg(feature = "__json")]
 pub mod json;
-pub mod layer;
-pub mod middleware;
-pub mod param;
 pub mod request;
 pub mod response;
-pub mod route;
+#[cfg(feature = "server")]
 pub mod server;
 
-pub(crate) mod service_fn;
-
-mod macros;
+pub(crate) mod utils;
 
 #[doc(hidden)]
 pub mod prelude {
@@ -28,11 +22,11 @@ pub mod prelude {
     pub use hyper;
     pub use volo::net::Address;
 
-    #[cfg(feature = "cookie")]
-    pub use crate::cookie::CookieJar;
-    #[cfg(any(feature = "serde_json", feature = "sonic_json"))]
+    pub use crate::extension::Extension;
+    #[cfg(feature = "__json")]
     pub use crate::json::Json;
-    pub use crate::{extension::Extension, param::Params, route::Router, server::Server};
+    #[cfg(feature = "server")]
+    pub use crate::server::prelude::*;
 }
 
 pub use self::prelude::*;
