+++ volo-http/src/response/mod.rs
@@ -1,8 +1,5 @@
-mod into_response;
-
 use hyper::body::Incoming;
 
-pub use self::into_response::IntoResponse;
 use crate::body::Body;
 
 pub type ServerResponse<B = Body> = http::Response<B>;
