+++ volo-http/src/server/mod.rs
@@ -20,11 +20,24 @@ use tokio::sync::Notify;
 use tracing::{info, trace};
 use volo::net::{conn::Conn, incoming::Incoming, Address, MakeIncoming};
 
-use crate::{
-    context::ServerContext,
-    request::ServerRequest,
-    response::{IntoResponse, ServerResponse},
-};
+use crate::{context::ServerContext, request::ServerRequest, response::ServerResponse};
+
+mod into_response;
+pub use self::into_response::IntoResponse;
+
+pub mod extract;
+mod handler;
+pub mod layer;
+pub mod middleware;
+pub mod param;
+pub mod route;
+
+#[doc(hidden)]
+pub mod prelude {
+    pub use super::{param::Params, route::Router, Server};
+    #[cfg(feature = "cookie")]
+    pub use crate::cookie::CookieJar;
+}
 
 /// This is unstable now and may be changed in the future.
 #[doc(hidden)]
