+++ volo-http/src/utils/macros.rs
@@ -42,5 +42,55 @@ macro_rules! all_the_tuples_no_last_special_case {
     };
 }
 
+macro_rules! impl_deref_and_deref_mut {
+    ($type:ty, $inner:ty, $pos:tt) => {
+        impl std::ops::Deref for $type {
+            type Target = $inner;
+
+            #[inline]
+            fn deref(&self) -> &Self::Target {
+                &self.$pos
+            }
+        }
+
+        impl std::ops::DerefMut for $type {
+            fn deref_mut(&mut self) -> &mut Self::Target {
+                &mut self.$pos
+            }
+        }
+    };
+}
+
+macro_rules! impl_getter {
+    ($name: ident, $type: ty, $($path: tt).+) => {
+        paste! {
+            #[inline]
+            pub fn $name(&self) -> &$type {
+                &self.$($path).+
+            }
+
+            #[inline]
+            pub fn [<$name _mut>](&mut self) -> &mut $type {
+                &mut self.$($path).+
+            }
+        }
+    };
+    ($name: ident, $type: ty) => {
+        paste! {
+            #[inline]
+            pub fn $name(&self) -> &$type {
+                &self.$name
+            }
+
+            #[inline]
+            pub fn [<$name _mut>](&mut self) -> &mut $type {
+                &mut self.$name
+            }
+        }
+    };
+}
+
 pub(crate) use all_the_tuples;
 pub(crate) use all_the_tuples_no_last_special_case;
+pub(crate) use impl_deref_and_deref_mut;
+pub(crate) use impl_getter;
