+++ volo-http/src/utils/mod.rs
@@ -1,0 +1,5 @@
+#![allow(unused)]
+pub mod macros;
+mod service_fn;
+
+pub use self::service_fn::{service_fn, Callback};
